refactor: Home에 있던 데이터 조회 코드 list로 옮김 및 필요없는 소스 삭제
